refactor(CO-3211): update backupsearch type imports to use specific paths for better clarity
